DownloadLinksPanel: fixed dialog bug Youtube: fixed DirectHTTP: added support for html redirects (highly experimental)
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@11295 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/gui/swing/jdgui/views/downloads/DownloadLinksPanel.java
@@ -539,7 +539,7 @@ public class DownloadLinksPanel extends SwitchPanel implements ActionListener, D
                     }
                     break;
                 case TableAction.DELETE: {
-                    if (JDFlags.hasSomeFlags(UserIO.getInstance().requestConfirmDialog(UserIO.DONT_SHOW_AGAIN | UserIO.DONT_SHOW_AGAIN_IGNORES_CANCEL, JDL.L("gui.downloadlist.delete", "Delete selected links?") + " (" + JDL.LF("gui.downloadlist.delete.size_packagev2", "%s links", selectedLinks.size()) + ")"), UserIO.RETURN_OK, UserIO.RETURN_DONT_SHOW_AGAIN)) {
+                    if (selectedLinks.size() > 0 && JDFlags.hasSomeFlags(UserIO.getInstance().requestConfirmDialog(UserIO.DONT_SHOW_AGAIN | UserIO.DONT_SHOW_AGAIN_IGNORES_CANCEL, JDL.L("gui.downloadlist.delete", "Delete selected links?") + " (" + JDL.LF("gui.downloadlist.delete.size_packagev2", "%s links", selectedLinks.size()) + ")"), UserIO.RETURN_OK, UserIO.RETURN_DONT_SHOW_AGAIN)) {
                         for (int i = 0; i < selectedLinks.size(); i++) {
                             selectedLinks.get(i).setEnabled(false);
                         }
@@ -551,17 +551,19 @@ public class DownloadLinksPanel extends SwitchPanel implements ActionListener, D
                     return;
                 }
                 case TableAction.DELETEFILE: {
-                    int counter = 0;
-                    for (DownloadLink tmp : selectedLinks) {
-                        if (tmp.existsFile()) counter++;
-                    }
-                    if (JDFlags.hasSomeFlags(UserIO.getInstance().requestConfirmDialog(UserIO.DONT_SHOW_AGAIN | UserIO.DONT_SHOW_AGAIN_IGNORES_CANCEL, JDL.L("gui.downloadlist.delete2", "Delete links from downloadlist and disk?") + " (" + JDL.LF("gui.downloadlist.delete.links", "%s links", selectedLinks.size()) + " / " + " " + JDL.LF("gui.downloadlist.delete.files", "%s files", counter) + ")"), UserIO.RETURN_OK, UserIO.RETURN_DONT_SHOW_AGAIN)) {
-                        for (int i = 0; i < selectedLinks.size(); i++) {
-                            selectedLinks.get(i).setEnabled(false);
+                    if (selectedLinks.size() > 0) {
+                        int counter = 0;
+                        for (DownloadLink tmp : selectedLinks) {
+                            if (tmp.existsFile()) counter++;
                         }
-                        for (int i = 0; i < selectedLinks.size(); i++) {
-                            selectedLinks.get(i).deleteFile(true, true);
-                            selectedLinks.get(i).getFilePackage().remove(selectedLinks.get(i));
+                        if (JDFlags.hasSomeFlags(UserIO.getInstance().requestConfirmDialog(UserIO.DONT_SHOW_AGAIN | UserIO.DONT_SHOW_AGAIN_IGNORES_CANCEL, JDL.L("gui.downloadlist.delete2", "Delete links from downloadlist and disk?") + " (" + JDL.LF("gui.downloadlist.delete.links", "%s links", selectedLinks.size()) + " / " + " " + JDL.LF("gui.downloadlist.delete.files", "%s files", counter) + ")"), UserIO.RETURN_OK, UserIO.RETURN_DONT_SHOW_AGAIN)) {
+                            for (int i = 0; i < selectedLinks.size(); i++) {
+                                selectedLinks.get(i).setEnabled(false);
+                            }
+                            for (int i = 0; i < selectedLinks.size(); i++) {
+                                selectedLinks.get(i).deleteFile(true, true);
+                                selectedLinks.get(i).getFilePackage().remove(selectedLinks.get(i));
+                            }
                         }
                     }
                     return;
